Handle possible undefined db

+++ src/test/testSetup.ts
@@ -33,7 +33,9 @@ const cleanDatabase = async function (): Promise<void> {
 	/// ////////////////////////////////////////////
 	/// ///////////////////////////////////////////
 	logger.debug('Cleaning databases')
-	await mongoose.connection.db.dropDatabase()
+	if (mongoose.connection.db !== undefined) {
+		await mongoose.connection.db.dropDatabase()
+	}
 }
 
 before(async function () {
